refactor: 팔로우 관련 패키지 정리 #58
