+++ server/src/main/java/com/hobbyt/domain/follow/repository/FollowRepository.java
@@ -1,10 +1,10 @@
-package com.hobbyt.domain.member.repository;
+package com.hobbyt.domain.follow.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.hobbyt.domain.member.entity.Follow;
+import com.hobbyt.domain.follow.entity.Follow;
 import com.hobbyt.domain.member.entity.Member;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
